apps/healthlib/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "healthlib"
version = "0.1.0"
description = "Personal library for interacting with Garmin and Strava APIs"
requires-python = ">=3.11"
dependencies = [
    "garminconnect>=0.2",
    "garth>=0.4",
    "requests>=2.28",
    "pyyaml>=6.0",
]

[project.scripts]
healthlib = "healthlib.__main__:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "mypy>=1.0",
    "ruff>=0.1",
    "types-requests",
    "types-PyYAML",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
]

[tool.mypy]
python_version = "3.11"
strict = true
